refactor informal breached after letter classification
